Adjust ShapeRenderer to avoid vectorial renderings be painted always on top. Fixes #323
